auth in lib
