+++ lib/auth.ts
@@ -1,0 +1,39 @@
+import NextAuth from "next-auth";
+import GithubProvider from "next-auth/providers/github";
+import CredentialProvider from "next-auth/providers/credentials";
+
+export const {
+  handlers: { GET, POST },
+  auth,
+} = NextAuth({
+  providers: [
+    GithubProvider({
+      clientId: process.env.GITHUB_ID ?? "",
+      clientSecret: process.env.GITHUB_SECRET ?? "",
+    }),
+    CredentialProvider({
+      credentials: {
+        email: {
+          label: "email",
+          type: "email",
+          placeholder: "example@gmail.com",
+        },
+      },
+      async authorize(credentials, req) {
+        const user = { id: "1", name: "John", email: credentials?.email };
+        if (user) {
+          // Any object returned will be saved in `user` property of the JWT
+          return user;
+        } else {
+          // If you return null then an error will be displayed advising the user to check their details.
+          return null;
+
+          // You can also Reject this callback with an Error thus the user will be sent to the error page with the error message as a query parameter
+        }
+      },
+    }),
+  ],
+  pages: {
+    signIn: "/", //sigin page
+  },
+});
